Update: Move OnboardingProgress to separate UI component
